Seaparate components into subsystems
