--- Assets/Scripts/ActiveRagdoll/RagdollState.cs
@@ -1,11 +1,0 @@
-namespace ActiveRagdoll
-{
-    public class RagdollState
-    {
-        public bool jumping;
-        public bool isJumping;
-        public bool inAir;
-        public bool punchingRight;
-        public bool punchingLeft;
-    }
-}
